requirements.txt
httpx
webdriver-manager
selenium
selenium-wire
cloudscraper
beautifulsoup4
rapidfuzz
PySide6==6.9.1
keyring==25.6.0
tinydb==4.8.2
openai==1.99.1
PyYAML==6.0.2
requests-toolbelt==1.0.0